+++ app/src/main/java/ru/netology/nework/repository/PostRepository.kt
@@ -1,12 +1,15 @@
 package ru.netology.nework.repository
 
 import ru.netology.nework.models.*
+import ru.netology.nework.models.post.Post
+import ru.netology.nework.models.post.PostCreateRequest
+import ru.netology.nework.models.user.User
 
 interface PostRepository {
-    suspend fun likeById(id: Int, likedByMe: Boolean): Post
-    suspend fun getUserById(id: Int): User
+    suspend fun likeById(id: Long, likedByMe: Boolean): Post
+    suspend fun getUserById(id: Long): User
     suspend fun save(post: PostCreateRequest)
     suspend fun saveWithAttachment(post: PostCreateRequest, upload: MediaUpload)
     suspend fun upload(upload: MediaUpload): Media
-    suspend fun removeById(id: Int)
+    suspend fun removeById(id: Long)
 }
